Both versions
The Bike Rental System is a simple command-line application that allows users to rent bikes from a bike shop. The program provides options to display the available stock, rent bikes by specifying the quantity, and exit the program. Also there is a twist in this repository. Here it contains two files, one is using 'try-except' block and another is normal, without using it. But the try-except one is more convinient. You use the one you like and convinient for you. If you encounter any issues or have suggestions for improvements, please open an issue or submit a pull request.